API/requirements.txt
flask
python-dotenv
requests
selenium
webdriver-manager
google-cloud-storage
google-cloud-run
google-cloud-iam
